// frontend/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  ai: typeof ai;
  auth: typeof auth;
  boards: typeof boards;
  email: typeof email;
  http: typeof http;
  images: typeof images;
  notifications: typeof notifications;
  payments: typeof payments;
  progressStore: typeof progressStore;
  promptAgent: typeof promptAgent;
  promptCategories: typeof promptCategories;
  promptExamples: typeof promptExamples;
  promptExamplesActions: typeof promptExamplesActions;
  promptFeedback: typeof promptFeedback;
  promptGenerator: typeof promptGenerator;
  promptImprover: typeof promptImprover;
  promptImproverActions: typeof promptImproverActions;
  promptRequests: typeof promptRequests;
  rag: typeof rag;
  rateLimits: typeof rateLimits;
  retrier: typeof retrier;
  seed: typeof seed;
  systemPrompts: typeof systemPrompts;
  users: typeof users;
  videos: typeof videos;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;